Move and rename stuff.
