+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/Spec.java
@@ -28,6 +28,7 @@ import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.cache.IndexedAttestationCache;
 import tech.pegasys.teku.spec.config.SpecConfig;
+import tech.pegasys.teku.spec.containers.ForkAndSpecMilestone;
 import tech.pegasys.teku.spec.datastructures.attestation.ValidateableAttestation;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockAndState;
@@ -51,8 +52,6 @@ import tech.pegasys.teku.spec.datastructures.state.beaconstate.MutableBeaconStat
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.common.BeaconStateInvariants;
 import tech.pegasys.teku.spec.datastructures.util.AttestationProcessingResult;
 import tech.pegasys.teku.spec.genesis.GenesisGenerator;
-import tech.pegasys.teku.spec.logic.ForkAndMilestone;
-import tech.pegasys.teku.spec.logic.Milestone;
 import tech.pegasys.teku.spec.logic.common.block.BlockProcessor;
 import tech.pegasys.teku.spec.logic.common.operations.validation.OperationInvalidReason;
 import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.BlockProcessingException;
@@ -128,21 +127,21 @@ public class Spec {
     return forkManifest;
   }
 
-  public Collection<ForkAndMilestone> getEnabledMilestones() {
+  public Collection<ForkAndSpecMilestone> getEnabledMilestones() {
     return forkManifest.getForkSchedule().stream()
-        .map(fork -> new ForkAndMilestone(fork, getMilestoneForFork(fork)))
+        .map(fork -> new ForkAndSpecMilestone(fork, getMilestoneForFork(fork)))
         .collect(toList());
   }
 
-  private Milestone getMilestoneForFork(final Fork fork) {
+  private SpecMilestone getMilestoneForFork(final Fork fork) {
     final SpecConfig specConfig = getSpecConfig(fork.getEpoch());
     if (fork.getCurrent_version().equals(specConfig.getGenesisForkVersion())) {
-      return Milestone.PHASE0;
+      return SpecMilestone.PHASE0;
     } else if (specConfig
         .toVersionAltair()
         .map(config -> fork.getCurrent_version().equals(config.getAltairForkVersion()))
         .orElse(false)) {
-      return Milestone.ALTAIR;
+      return SpecMilestone.ALTAIR;
     } else {
       throw new UnsupportedOperationException("Unsupported fork scheduled" + fork);
     }
